Moved AboutPage elements in LoginPage
-AboutPage is now imported into LoginPage instead
-Updated AboutPage with adjustments
